requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
supabase==1.0.0
motor==4.1.0
pdfplumber==0.8.1
python-docx==0.8.11
requests==2.31.0
pydantic==1.10.11
python-dotenv==1.0.0
